Latest updates for first version of ssm
- migrate from pip to pipenv(2018.11.26)
- don't create system user if this user already exist or SYSTEM_EMAIL/SYSTEM_PASSWORD environment variables were not defined
- add rejected status for absences, better permissions check for non-staff absence POST/PUT/PATCH requests
- add IsStaff permission
- skills endpoint to return paginated list of skills and list of users who own these skills
- allow to update skills in users API endpoint
- add missed IsAuthenticated permissions for views

+++ dockerX/ssm/systemuser.py
@@ -5,7 +5,8 @@ import django; django.setup()
 from django.db import transaction
 models = importlib.import_module("%s.users.models" % os.environ['PROJECT'])
 
-if not models.User.objects.filter(email=os.environ['SYSTEM_EMAIL']).exists():
-    with transaction.atomic():
-        models.User.objects.filter(email=os.environ['SYSTEM_EMAIL']).delete()
-        models.User.objects.create_superuser(os.environ['SYSTEM_EMAIL'], os.environ['SYSTEM_PASSWORD'])
+if os.environ['SYSTEM_EMAIL'] and os.environ['SYSTEM_PASSWORD']:
+    if not models.User.objects.filter(email=os.environ['SYSTEM_EMAIL']).exists():
+        with transaction.atomic():
+            models.User.objects.filter(email=os.environ['SYSTEM_EMAIL']).delete()
+            models.User.objects.create_superuser(os.environ['SYSTEM_EMAIL'], os.environ['SYSTEM_PASSWORD'])
